requirements.txt
# EmpathAI Backend Requirements
# Python 3.11+ compatible

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Machine Learning
torch==2.5.1

transformers==4.35.0
sentence-transformers==2.2.2

# Google Gemini API for response generation
google-generativeai==0.3.1

# Additional utilities
numpy==1.24.3

